Update array access in eventSummary.js
Changed the way of accessing the first element in an event array from using .first() method to direct index access. This was done for readability and consistency purposes as it is a more common approach to access elements in an array in Javascript. Moreover, this should provide a small performance improvement as direct array index access is generally faster than invoking an array method.